Extract atomic_directory module from common.
